Revert "LPS-76837 Move to taglib module"
This reverts commit ce3138e14adad6714da00380a8b9ca7f2dd515cb.
