+++ modules/apps/foundation/text-localizer/text-localizer-address/src/main/java/com/liferay/text/localizer/address/USAddressTextLocalizer.java
@@ -12,15 +12,14 @@
  * details.
  */
 
-package com.liferay.text.localizer.taglib.internal;
+package com.liferay.text.localizer.address;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Address;
 import com.liferay.portal.kernel.util.Html;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.text.localizer.address.AddressTextLocalizer;
-import com.liferay.text.localizer.taglib.internal.util.AddressUtil;
+import com.liferay.text.localizer.address.util.AddressUtil;
 
 import java.util.Optional;
