+++ modules/apps/foundation/text-localizer/text-localizer-address/src/main/java/com/liferay/text/localizer/address/util/AddressTextLocalizerUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.text.localizer.taglib.internal.util;
+package com.liferay.text.localizer.address.util;
 
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
